spidermonkey 128.3.0
This updates `spidermonkey` to the latest ESR version.

The `livecheck` block is incorrectly returning 115.16.0 as the latest
version, as there haven't been release notes for any versions after
that point. This updates the `livecheck` block to check the
redirecting download URL for Firefox ESR assets, which will redirect
to a file that includes the ESR version that we can check.

What the URL redirects to is determined by query string parameters
(e.g. `os=osx` for the latest macOS dmg) but unfortunately there
doesn't appear to be an option to fetch the latest source tarball.
I omitted `os` and `lang` parameters, as the URL will fail an audit
if they're included: `https://download.mozilla.org/?product=firefox-
esr-latest-ssl&os=osx&lang=en-US looks like a binary package, not a
source archive; homebrew/core is source-only.` It's not ideal but it
works and we can always revisit it if we run into issues in the
future.